Use polling instead of the Listen gem
Listen uses too much memory, because it watches the whole directory recursively.
Since we store the attachments and per-client configuration files below the same
directory, this makes it setup watches for every attachment and log file, even if
they'll end up being discarded in the end.

Since the number of configuration files is small, polling is a good enough solution
for now.